Add files via upload
The Tic Tac Toe game is also known as “Noughts and crosses” game. It is one of the most popular two-player pen-and-paper games.

This two-player pen-and-paper game has extremely ancient roots and is popular in many countries.

The game’s rules are straightforward and well-known.

Two players, X and O, will play this game by alternately marking the squares in a 3X3 grid.

The game is won by the person who successfully arranges three of their marks in a row that is either horizontal, vertical, or diagonal.

Only one symbol may be placed by each player at every turn, after which the turn is passed to the other player.